+++ packages/core/src/sagas.js
@@ -4,16 +4,19 @@ import clearSuggestedActionsOnPostActivitySaga from './sagas/clearSuggestedActio
 import connectionStatusUpdateSaga from './sagas/connectionStatusUpdateSaga';
 import connectSaga from './sagas/connectSaga';
 import incomingActivitySaga from './sagas/incomingActivitySaga';
-import incomingTypingSaga from './sagas/incomingTypingSaga';
-import markActivityForSpeakSaga from './sagas/markActivityForSpeakSaga';
+import markActivityForSpeakOnIncomingActivityFromOthersSaga from './sagas/markActivityForSpeakOnIncomingActivityFromOthersSaga';
+import markAllAsSpokenOnStopSpeakActivitySaga from './sagas/markAllAsSpokenOnStopSpeakActivitySaga';
 import postActivitySaga from './sagas/postActivitySaga';
+import removeIncomingTypingAfterIntervalSaga from './sagas/removeIncomingTypingAfterIntervalSaga';
+import sendEventToPostActivitySaga from './sagas/sendEventToPostActivitySaga';
 import sendFilesToPostActivitySaga from './sagas/sendFilesToPostActivitySaga';
 import sendMessageToPostActivitySaga from './sagas/sendMessageToPostActivitySaga';
 import sendPostBackToPostActivitySaga from './sagas/sendPostBackToPostActivitySaga';
 import sendTypingOnSetSendBoxSaga from './sagas/sendTypingOnSetSendBoxSaga';
 import startDictateAfterSpeakActivitySaga from './sagas/startDictateAfterSpeakActivitySaga';
-import stopDictateOnCardAction from './sagas/stopDictateOnCardAction';
-import stopSpeakActivityOnInputSaga from './sagas/stopSpeakActivityOnInputSaga';
+import startSpeakActivityOnPostActivitySaga from './sagas/startSpeakActivityOnPostActivitySaga';
+import stopDictateOnCardActionSaga from './sagas/stopDictateOnCardActionSaga';
+import stopSpeakingActivityOnInputSaga from './sagas/stopSpeakingActivityOnInputSaga';
 import submitSendBoxSaga from './sagas/submitSendBoxSaga';
 
 export default function* () {
@@ -21,15 +24,18 @@ export default function* () {
   yield fork(connectionStatusUpdateSaga);
   yield fork(connectSaga);
   yield fork(incomingActivitySaga);
-  yield fork(incomingTypingSaga);
-  yield fork(markActivityForSpeakSaga);
+  yield fork(markActivityForSpeakOnIncomingActivityFromOthersSaga);
+  yield fork(markAllAsSpokenOnStopSpeakActivitySaga);
   yield fork(postActivitySaga);
+  yield fork(removeIncomingTypingAfterIntervalSaga);
+  yield fork(sendEventToPostActivitySaga);
   yield fork(sendFilesToPostActivitySaga);
   yield fork(sendMessageToPostActivitySaga);
   yield fork(sendPostBackToPostActivitySaga);
   yield fork(sendTypingOnSetSendBoxSaga);
   yield fork(startDictateAfterSpeakActivitySaga);
-  yield fork(stopDictateOnCardAction);
-  yield fork(stopSpeakActivityOnInputSaga);
+  yield fork(startSpeakActivityOnPostActivitySaga);
+  yield fork(stopDictateOnCardActionSaga);
+  yield fork(stopSpeakingActivityOnInputSaga);
   yield fork(submitSendBoxSaga);
 }