+++ packages/core/src/sagas/effects/whileConnected.js
@@ -5,13 +5,13 @@ import {
   take
 } from 'redux-saga/effects';
 
-import { CONNECT_FULFILLED } from '../../actions/connect';
+import { CONNECT_FULFILLING } from '../../actions/connect';
 import { DISCONNECT_FULFILLED } from '../../actions/disconnect';
 
 export default function (fn) {
   return call(function* () {
     for (;;) {
-      const { meta: { userID }, payload: { directLine } } = yield take(CONNECT_FULFILLED);
+      const { meta: { userID }, payload: { directLine } } = yield take(CONNECT_FULFILLING);
       const task = yield fork(fn, directLine, userID);
 
       yield take(DISCONNECT_FULFILLED);