+++ packages/core/src/sagas/effects/whileSpeakIncomingActivity.js
@@ -1,0 +1,22 @@
+import {
+  call,
+  cancel,
+  fork,
+  take
+} from 'redux-saga/effects';
+
+import { START_SPEAKING_ACTIVITY } from '../../actions/startSpeakingActivity';
+import { STOP_SPEAKING_ACTIVITY } from '../../actions/stopSpeakingActivity';
+
+export default function (fn) {
+  return call(function* () {
+    for (;;) {
+      yield take(START_SPEAKING_ACTIVITY);
+
+      const task = yield fork(fn);
+
+      yield take(STOP_SPEAKING_ACTIVITY);
+      yield cancel(task);
+    }
+  });
+}